int fix

+++ slp-queries/bm-text-to-lz77/include/utils/bm_compression.hpp
@@ -109,7 +109,7 @@ public:
                 // current_hash = stringToHash(curr);
 
                 char_t* curr_str = allocate_array<char_t>(block_size);
-                for(int i = position; i < position + block_size; i++) {
+                for(uint64_t i = position; i < position + block_size; i++) {
                     curr_str[i - position] = (text[i]);
                 }
                 current_hash = hash_string<char_t>(curr_str, block_size);
@@ -348,7 +348,7 @@ public:
         cout << "Num Blocks: " << blocks.size() << endl;
         cout << "Reduction: " << 100 - (double)(parsing.size() * 16) * 100 / (strlen(reinterpret_cast<const char*>(text))) << '%' << endl;
 
-        for(int i = 0; i < parsing.size(); ++i) {
+        for(uint64_t i = 0; i < parsing.size(); ++i) {
             // cout << parsing[i].first << ' ' << parsing[i].second << endl;
         }
     }
